Darwin: rework syscall arg handling
There were numerous issues, particularly with syscall SYS_syscall
and the scalar tests. Only tested on amd64, I don't have access to
an OSX 10.12 or earlier machine to do any testing.

The key elements in VG_(client_syscall) are, when the PRE wrapper
does not handle the syscall

1. getSyscallArgsFromGuestState
2. getSyscallArgLayout
3. call PRE wrapper

either blocking

4. putSyscallArgsIntoGuestState
5. do_syscall_for_client

or non-blocking

6. VG_(do_syscall)

What has changed
----------------
struct SyscallArgs now like FreeBSD, stores both the
canonical_sysno (which is the final syscall number, never
SYS_syscall) and the original_sysno (which can be SYS_syscall).
These get set in getSyscallArgsFromGuestState which has
changed accordingly.

getSyscallArgsFromGuestState has changed to always get the layout
coresponding to the canonical layout. This was the cause of the
problems with the scalar tests - the arguments were effectively
out by one.

Many subsequent changes to the scalar test, filter and expected.
The scalar test still isn't quite right. __NR_kevent_qos is missing
a scalar error from argument 6. That's the one with special case
handling because it is a register argument for a normal syscall
but a stack argument for syscall SYS_syscall.