s390x: Support FPC rounding mode 0b111
When get_bfp_rounding_mode_from_fpc was added Irrm_PREPARE_SHORTER was
not available. The rounding mode 0x111 in the FPC was mapped to Irrm_NEAREST
with fingers crossed.
When Irrm_PREPARE_SHORTER was added later on it was forgotten to adjust this
function. Now fixed. Comments and bfp-convert testcase  adjusted.