Massif regtest: filter ignore functions, part 1
This is mainly for Darwin which has numerous leaks that
need to be ignored. This filter removes all --ignore-fn options
from the "out" expecteds. This should allow adding endless
functions to ignore in the vgtest files without having to
also update the out.exp files as well.

Part 2 will be for the verbose output which adds ignore files
to the stderr.exp references.