s390x: Simplify single multiplications
So far the single-multiplication instructions are implemented by taking
the low half of a widening multiply.  This pattern is then recognized by
the instruction selector, which combines it to a single multiplication
again.

This is unnecessarily complicated, since single-multiply operations can be
expressed directly by Iop_Mul32 and Iop_Mul64 instead.  So do this,
simplifying the code generation.

Also, since Iop_Mul32 and Iop_Mul64 haven't really been used before, the
instruction selector generates sub-optimal code for them.  Fix that.