Darwin: also clean up client stack creation
Make stringbase and strtab be based off clstack_end + 1
rather than clstack_end. clstack_end is the address of the
last byte of the stack, not one past the end of the stack.
That means that it is not word aligned. Adding 1 makes
the calculation word aligned.

Depending on the length of the string table with rounding
it could happen that the executable_path pointer would
overwrite the start of argv[0] in the string table.

I think that that there are supposed to be 0-16 bytes between
the end of the pointers and the string table (or is that
ELF?).