Darwin configure: updates to SDK checking and not yet suppoorted Darwin versions
Merged from Louis Brunner

I was just adding support for OSX 10.13 but that is just making more
work (avoiding parts of merges that will probably need to be added back later).

This should allow me to merge code that uses macros for OSX/macOS versions
which don't yet have support in the repo.