Darwin regtest: numerous bits of Darwin 18 cleanup
Add a few more ignore-fn options to the massif tests for the
extra leaks that are in Darwin 18.

Add some suppressions to a couple of memcheck tests that use
num-callers of 2 or 3. The shorter callsteack means some of the
many Darwin default suppressions don't match which was causing
these tests to fail.

Add more DRD suppressions for getaddrinfo. There seems to be a never
ending amount of conflicts in dyld.